refactor: [헥사고날 아키텍쳐] application.port in, out에 해당하는 요소 구성
