+++ src/main/java/com/lgmpjt/websnsspringboot/adapter/in/rest/BoardController.java
@@ -1,7 +1,7 @@
 package com.lgmpjt.websnsspringboot.adapter.in.rest;
 
-import com.lgmpjt.websnsspringboot.domain.board.data.BoardCreateDto;
-import com.lgmpjt.websnsspringboot.domain.board.data.BoardDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardCreateDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardDto;
 import com.lgmpjt.websnsspringboot.domain.board.service.BoardService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
