+++ src/main/java/com/lgmpjt/websnsspringboot/adapter/in/rest/LikeController.java
@@ -1,6 +1,6 @@
 package com.lgmpjt.websnsspringboot.adapter.in.rest;
 
-import com.lgmpjt.websnsspringboot.domain.board.data.BoardDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardDto;
 import com.lgmpjt.websnsspringboot.domain.like.service.LikeService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
