+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/in/LikeCommandUseCase.java
@@ -1,0 +1,4 @@
+package com.lgmpjt.websnsspringboot.application.port.in;
+
+public interface LikeCommandUseCase {
+}
