+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/in/dto/BoardCreateDto.java
@@ -1,6 +1,5 @@
-package com.lgmpjt.websnsspringboot.domain.board.data;
+package com.lgmpjt.websnsspringboot.application.port.in.dto;
 
-import com.lgmpjt.websnsspringboot.domain.user.data.UserDto;
 import lombok.*;
 import org.springframework.util.Assert;
 
@@ -11,24 +10,18 @@ import java.time.LocalDateTime;
 @Setter
 @Builder
 @NoArgsConstructor
-@AllArgsConstructor
-public class BoardDto {
-	private Long boardSeq;
+public class BoardCreateDto {
 	private UserDto user;
 	private String content;
 	private String boardImage;
 	private LocalDateTime createdDate;
-	private LocalDateTime lastModifiedDate;
 
-	public BoardDto(final Long boardSeq, final UserDto user, final String content, final String boardImage, final LocalDateTime createdDate) {
-		Assert.notNull(boardSeq, "게시물 SEQ 정보는 필수입니다.");
+	public BoardCreateDto(UserDto user, String content, String boardImage, LocalDateTime createdDate) {
 		Assert.notNull(user, "게시물을 작성한 유저 정보는 필수입니다.");
 		Assert.hasText(boardImage, "게시물의 이미지는 필수입니다.");
-		this.boardSeq = boardSeq;
 		this.user = user;
 		this.content = content;
 		this.boardImage = boardImage;
 		this.createdDate = createdDate;
-		this.lastModifiedDate = LocalDateTime.now();
 	}
 }
