+++ src/main/java/com/lgmpjt/websnsspringboot/mapper/BoardMapper.java
@@ -1,9 +1,9 @@
-package com.lgmpjt.websnsspringboot.mapper.board;
+package com.lgmpjt.websnsspringboot.mapper;
 
-import com.lgmpjt.websnsspringboot.domain.board.data.BoardCreateDto;
-import com.lgmpjt.websnsspringboot.domain.board.data.BoardDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardCreateDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.BoardDto;
 import com.lgmpjt.websnsspringboot.adapter.out.entity.Board;
-import com.lgmpjt.websnsspringboot.domain.user.data.UserDto;
+import com.lgmpjt.websnsspringboot.application.port.in.dto.UserDto;
 import com.lgmpjt.websnsspringboot.adapter.out.entity.User;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
